+++ supervisely/nn/benchmark/utils/__init__.py
@@ -1,9 +1,9 @@
 # isort: skip_file
-from supervisely.nn.benchmark.utils.coco.calculate_metrics import calculate_metrics
-from supervisely.nn.benchmark.utils.coco.metric_provider import MetricProvider
-from supervisely.nn.benchmark.utils.coco.sly2coco import sly2coco
-from supervisely.nn.benchmark.utils.coco.utlis import read_coco_datasets
-from supervisely.nn.benchmark.utils.utils import try_set_conf_auto
+from supervisely.nn.benchmark.utils.detection.calculate_metrics import calculate_metrics
+from supervisely.nn.benchmark.utils.detection.metric_provider import MetricProvider
+from supervisely.nn.benchmark.utils.detection.sly2coco import sly2coco
+from supervisely.nn.benchmark.utils.detection.utlis import read_coco_datasets
+from supervisely.nn.benchmark.utils.detection.utlis import try_set_conf_auto
 
 from supervisely.nn.benchmark.utils.semantic_segmentation.calculate_metrics import (
     calculate_metrics as calculate_semsegm_metrics,
