--- supervisely/nn/benchmark/utils/coco/__init__.py
@@ -1,2 +1,0 @@
-from supervisely.nn.benchmark.utils.coco.calculate_metrics import calculate_metrics
-from supervisely.nn.benchmark.utils.coco.metric_provider import MetricProvider
