+++ supervisely/nn/benchmark/utils/detection/__init__.py
@@ -1,0 +1,2 @@
+from supervisely.nn.benchmark.utils.detection.calculate_metrics import calculate_metrics
+from supervisely.nn.benchmark.utils.detection.metric_provider import MetricProvider
