+++ supervisely/nn/benchmark/utils/detection/metric_provider.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 import numpy as np
 import pandas as pd
 
-from supervisely.nn.benchmark.utils.coco import metrics
+from supervisely.nn.benchmark.utils.detection import metrics
 
 METRIC_NAMES = {
     "mAP": "mAP",
