+++ supervisely/nn/benchmark/utils/detection/utlis.py
@@ -1,4 +1,15 @@
 from supervisely.io.json import load_json_file
+from supervisely.nn.inference import SessionJSON
+
+
+def try_set_conf_auto(session: SessionJSON, conf: float):
+    conf_names = ["conf", "confidence", "confidence_threshold", "confidence_thresh"]
+    default = session.get_default_inference_settings()
+    for name in conf_names:
+        if name in default:
+            session.inference_settings[name] = conf
+            return True
+    return False
 
 
 def read_coco_datasets(cocoGt_json, cocoDt_json):
