+++ supervisely/nn/benchmark/visualization/vis_click_data.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from collections import defaultdict
 
-from supervisely.nn.benchmark.utils.coco.metric_provider import (
+from supervisely.nn.benchmark.utils.detection.metric_provider import (
     MetricProvider,
 )
